Update dice.js
OK. The NaN error was from the parseINT meeting a string. I don't understand what the block of code below each case set does (line 450 area), but it seems to be identical between blocks so it should be for "gm" too. There should be something to cull maths from a gm roll, 2gm could output 2 different results, but 2gm+1 is meaningless. I was going to copy the swd filter, but /roll swd dosen't work in slack anymore, so this is not a fully viable change.